Added automation and safety checks
Code now checks whether the remote control is detected and the arming switch so the drone can be kept safe. Also, CH6 is now used as an automation selector. When at the default value, the channel values taken from the remote-control are just passed through to the drone, the Arduino is acting as a relay. When this selector value changes however, the drone enters its automated mode. Currently this consists of setting channel values for selected periods of time, enabling a rough demo routine. This will be later expanded to include localisation and a PID controller.