ESP crashing
Speed controllers are implemented and velocity data is reported to ROS to be plotted. Occasionally (seemingly only when ESP is connected to FCU), it crashes with "Load/Store prohibited", suggested pointers to corrupt/uninitialised data. This occurs at Wire::beginTransmission, seemingly to do with the IMU communication. Currently attempting to reduce the number of global variables, especially arrays to see if this makes any difference.